Update patch set 1
Patch Set 1:

The light HALs will be converted to hidl and moved to hardware/lineage/interfaces. Then we can kill the BOARD_USES_CYANOGEN_HARDWARE, it's used only for the light HALs. 

After that, only lineagehw will remain in this repo, which we plan to completely refactor the way that it works in the (near?) future (convert it to hidl interfaces)

Patch-set: 1
Reviewer: Michael Bestas <4550@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0